Extract versioning and add wasmer module version to folder name
