+++ packages/vm/src/modules/mod.rs
@@ -2,7 +2,9 @@ mod file_system_cache;
 mod in_memory_cache;
 mod pinned_memory_cache;
 mod sized_module;
+mod versioning;
 
 pub use file_system_cache::FileSystemCache;
 pub use in_memory_cache::InMemoryCache;
 pub use pinned_memory_cache::PinnedMemoryCache;
+pub use versioning::current_wasmer_module_version;
